Enhance documentation and configuration for Rybbit platform
- Updated the introduction section in the documentation to provide a clearer overview of Rybbit as a powerful open-source web and product analytics platform.
- Added remote image patterns in the Next.js configuration to support images from Twitter and UI avatars.
- Introduced new CSS animations for marquee effects to improve visual dynamics across the application.
- Refactored the Tweet component to streamline rendering and improve overall performance.
