+++ docs-v2/content/docs/changing-domains.mdx
@@ -1,0 +1,17 @@
+---
+title: Changing domains
+description: Learn how to update the domain of your site in Rybbit
+---
+
+import Image from "next/image";
+
+# Changing domains
+
+You can update the domain of your site by going to the settings page on your site's Rybbit dashboard.
+
+Changing domains will not affect any existing data.
+
+Currently we only support one domain per site.
+
+<br/>
+<Image src="/settings.jpg" alt="Public Site" width={500} height={500} />
