+++ docs-v2/content/docs/definitions.mdx
@@ -1,0 +1,115 @@
+---
+title: Definitions
+description: Understanding Rybbit's metrics and data points
+---
+
+# Definitions
+
+## Aggregate Metrics
+
+### Unique Users
+The total number of distinct individuals who visited your website, identified by a unique user ID. Each user is counted only once regardless of how many times they visit.
+
+### Sessions
+A series of user interactions with your website within a given time frame. A session ends after 30 minutes of inactivity from the same user ID. Multiple pageviews can occur within one session.
+
+### Pageviews
+The total number of times visitors have viewed pages on your website. Each time a page is loaded or reloaded counts as a pageview.
+
+### Pages Per Session
+The average number of pages viewed during a session. Calculated by dividing the total number of pageviews by the total number of sessions.
+
+### Bounce Rate
+The percentage of sessions where a user leaves your site after viewing only one page without any interaction. A high bounce rate might indicate that landing pages aren't relevant to visitors.
+
+### Session Duration
+The average length of time users spend on your site during a session. Longer sessions typically indicate more engaged visitors.
+
+## Individual Datapoints
+
+### User ID
+A unique identifier generated from a hash of the visitor's IP address and browser user-agent. This anonymized ID helps track unique visitors without storing personally identifiable information.
+
+### Hostname
+The domain name of the website being visited (e.g., example.com).
+
+### Pathname
+The path portion of the URL that follows the hostname (e.g., /blog/article).
+
+### Entry Page
+The first page viewed by a user during a session. Identifies which pages commonly serve as entry points to your site.
+
+### Exit Page
+The last page viewed by a user before ending their session. Helps identify where users tend to leave your site.
+
+### Querystring
+The portion of a URL that follows a question mark, containing parameters (e.g., ?utm_source=newsletter).
+
+### Page Title
+The title of the webpage as defined in the HTML `<title>` tag.
+
+### Referrer
+The URL that a user was on before coming to your site. Shows how users are discovering your content.
+
+### UTM Source
+Identifies which website, search engine, or other source is sending traffic to your site (e.g., google, newsletter, twitter).
+
+### UTM Medium
+Identifies the marketing medium (e.g., cpc, email, social).
+
+### UTM Term
+Identifies paid search keywords. Used primarily for paid search campaigns.
+
+### UTM Campaign
+Identifies the specific campaign or promotion that brought users to your site.
+
+### UTM Content
+Identifies what specifically was clicked to bring the visitor to your site (e.g., banner ad, text link).
+
+### Channel
+A categorization of traffic sources into standard groups such as direct, search, social, email, or referral.
+
+### Browser
+The web browser used by the visitor (e.g., Chrome, Firefox, Safari).
+
+### Browser Version
+The specific version of the browser being used.
+
+### Operating System
+The operating system used by the visitor (e.g., Windows, macOS, iOS, Android).
+
+### Operating System Version
+The specific version of the operating system being used.
+
+### Language
+The preferred language setting of the visitor's browser.
+
+### Country
+The country where the user is located, based on IP geolocation.
+
+### Region
+The state or region where the user is located, based on IP geolocation.
+
+### City
+The city where the user is located, based on IP geolocation.
+
+### Latitude
+The geographic latitude coordinate of the user's approximate location.
+
+### Longitude
+The geographic longitude coordinate of the user's approximate location.
+
+### Screen Width
+The width of the visitor's screen or device viewport in pixels.
+
+### Screen Height
+The height of the visitor's screen or device viewport in pixels.
+
+### Device Type
+The type of device being used (e.g., desktop, tablet, mobile).
+
+### Event Name
+The name of a custom event triggered by user interaction (e.g., "button_click", "form_submit").
+
+### Event Properties
+Additional contextual data associated with custom events, stored as key-value pairs. Used to provide more detail about specific interactions.
