+++ docs-v2/content/docs/deleting-sites.mdx
@@ -1,0 +1,13 @@
+---
+title: Deleting sites
+description: Learn how to delete a site and all its associated data from Rybbit
+---
+
+import Image from "next/image";
+
+# Deleting sites
+
+You can delete a site by going to the settings page on the site's dashboard and clicking the "Delete Site" button. This will delete all data associated with the site.
+
+<br/>
+<Image src="/settings.jpg" alt="Public Site" width={500} height={500} />
