+++ docs-v2/content/docs/enhanced-privacy.mdx
@@ -1,0 +1,28 @@
+---
+title: Enhanced Privacy
+description: Learn about Rybbit's enhanced privacy features including user ID salting for GDPR compliance
+---
+
+import Image from "next/image";
+import { Callout } from 'fumadocs-ui/components/callout';
+
+# Enhanced Privacy
+
+{/* <Callout type="info">
+  If strict GDPR/CCPA compliance is not important to you, leave this option disabled as it makes some parts of the analytics less accurate.
+</Callout> */}
+
+Rybbit stores user IDs as a hashed combination of IP + user agent. Doing this allows us to identify users across many sessions on the same device, but can be seen as storing personally identifiable data under GDPR.
+
+You can enable enhanced privacy by going to the settings page of your website dashboard and enabling the "User ID Salting" option.
+
+When you turn this one, Rybbit will forget about unique users every day. If a user visits your site 10 times in one day, Rybbit will count that as one user with 10 sessions. But if he visits once a day for 10 days, he will be seen as a new user every day.
+
+<br/>
+<Image src="/settings.jpg" alt="Public Site" width={500} height={500} />
+
+### Downsides of enabling user ID salting:
+
+- You will no longer see user session history beyond 24 hours.
+- Unique user count over a period longer than 24 hours will be inflated.
+- The user retention page will effectively be useless.
