+++ docs-v2/content/docs/guides/meta.json
@@ -1,0 +1,44 @@
+{
+  "title": "Integration Guides",
+  "pages": [
+    "angular",
+    "astro",
+    "bubble",
+    "docusaurus",
+    "drupal",
+    "framer",
+    "google-tag-manager",
+    "hugo",
+    "jekyll",
+    "laravel",
+    {
+      "title": "React",
+      "pages": [
+        "react/vite-cra",
+        "react/next-js",
+        "react/gatsby",
+        "react/remix"
+      ]
+    },
+    "shopify",
+    "squarespace",
+    {
+      "title": "Svelte",
+      "pages": [
+        "svelte/vite",
+        "svelte/sveltekit"
+      ]
+    },
+    {
+      "title": "Vue",
+      "pages": [
+        "vue/vite",
+        "vue/nuxt"
+      ]
+    },
+    "webflow",
+    "wix",
+    "woocommerce",
+    "wordpress"
+  ]
+}
