+++ docs-v2/content/docs/guides/wix.mdx
@@ -1,0 +1,70 @@
+---
+title: "Wix"
+description: "Integrate Rybbit Analytics with your Wix website using custom code snippets."
+---
+
+import { Callout } from "fumadocs-ui/components/callout";
+
+# Wix
+
+Integrating Rybbit Analytics with your Wix website involves adding a custom code snippet to your site's tracking tools.
+
+## 1. Get Your Tracking Script
+
+First, you'll need your Rybbit tracking script. You can find this in your Rybbit dashboard under **Site Settings > Tracking Code**. It will look something like this:
+
+```html
+<script async defer src="https://app.rybbit.io/api/script.js" data-site-id="YOUR_SITE_ID"></script>
+```
+
+Replace `YOUR_SITE_ID` with your actual Site ID from your Rybbit dashboard.
+
+## 2. Go to Wix Dashboard
+
+Log in to your Wix account and select the site you want to track.
+
+From your site's dashboard, navigate to:
+
+- **Settings** (usually in the left-hand sidebar)
+
+## 3. Access Tracking Tools / Custom Code
+
+In the Settings menu, look for a section related to tracking or advanced settings. This is often called:
+
+- **Tracking Tools & Analytics** (under "Marketing & SEO" or a similar category)
+- Or, **Custom Code** (sometimes under "Advanced Settings")
+
+<Callout type="info">
+  To add custom code, you'll need a Wix Premium plan and a connected domain. This feature is typically not available on free Wix plans.
+</Callout>
+
+## 4. Add New Custom Code
+
+Once in the Tracking Tools or Custom Code section:
+
+- Click on **+ New Tool** or **+ Add Custom Code**.
+- Choose **Custom** from the list of options if prompted.
+
+## 5. Configure the Custom Code
+
+You'll be presented with a form to add your code:
+
+- **Paste your Rybbit tracking script** into the main code snippet box.
+- **Name your code:** Give it a recognizable name, like "Rybbit Analytics".
+- **Add Code to Pages:** Select **All pages** to track your entire site.
+  - You might have an option to **Load code once** or **Load code on each new page**. Choose the option that best reflects loading the script on every page view.
+- **Place Code in:** Choose **Body - end**. This is generally recommended for analytics scripts to avoid impacting perceived page load speed. Some Wix versions might offer "Head" or "Body - start" as well. "Body - end" is usually the best choice.
+
+## 6. Apply Changes
+
+Click **Apply** or **Save** to add the custom code to your site.
+
+## 7. Verify Integration
+
+After applying the changes, Rybbit Analytics should start tracking visitors on your Wix site.
+
+- Open your Wix website in a new browser tab or incognito window.
+- Navigate through a few pages.
+- Check your Rybbit dashboard for incoming data. It might take a few minutes for the first events to appear.
+
+That's it! Rybbit Analytics is now integrated with your Wix website.
