+++ docs-v2/content/docs/index.mdx
@@ -1,13 +1,64 @@
 ---
-title: Hello World
-description: Your first document
+title: Introduction
+description: Powerful open source web and product analytics platform
 ---
 
-Welcome to the docs! You can start writing documents in `/content/docs`.
+# Introduction
 
-## What is Next?
+**Rybbit** is a powerful open source web and products analytics platform. Check out our [live demo](https://demo.rybbit.io/1) using a real production site that sends over **15,000,000** events a month.
+
+## Key Features
+
+- Extremely comprehensive prebuilt web analytics dashboards
+- Cookieless and privacy-friendly
+- Ability to make your dashboard public
+- Support for organizations
+- Cool visualizations
+- Advanced product analytics like funnels, user retention, user journeys, and custom reports
+- Quick setup on any website within a few minutes
+
+## Getting Started
+
+Choose the option that's right for you:
+
+- **Cloud Version:** Sign up for the hosted version at [rybbit.io/signup](https://app.rybbit.io/signup). Get up and running in minutes.
+- **Self-Host:** Follow our [Self-Hosting Guide](/docs/self-hosting) to set up Rybbit on your own server.
+
+And then learn how to add the [Tracking Script](/docs/script) to your website to start collecting data.
+
+## Integration Guides
 
 <Cards>
-  <Card title="Learn more about Next.js" href="https://nextjs.org/docs" />
-  <Card title="Learn more about Fumadocs" href="https://fumadocs.vercel.app" />
+  <Card title="Angular" href="/docs/guides/angular" />
+  <Card title="Astro" href="/docs/guides/astro" />
+  <Card title="Bubble" href="/docs/guides/bubble" />
+  <Card title="Docusaurus" href="/docs/guides/docusaurus" />
+  <Card title="Drupal" href="/docs/guides/drupal" />
+  <Card title="Framer" href="/docs/guides/framer" />
+  <Card title="Google Tag Manager" href="/docs/guides/google-tag-manager" />
+  <Card title="Hugo" href="/docs/guides/hugo" />
+  <Card title="Jekyll" href="/docs/guides/jekyll" />
+  <Card title="Laravel" href="/docs/guides/laravel" />
+  <Card title="React" href="/docs/guides/react/vite-cra" />
+  <Card title="Next.js" href="/docs/guides/react/next-js" />
+  <Card title="Gatsby" href="/docs/guides/react/gatsby" />
+  <Card title="Remix" href="/docs/guides/react/remix" />
+  <Card title="Shopify" href="/docs/guides/shopify" />
+  <Card title="Squarespace" href="/docs/guides/squarespace" />
+  <Card title="Svelte" href="/docs/guides/svelte/vite" />
+  <Card title="SvelteKit" href="/docs/guides/svelte/sveltekit" />
+  <Card title="Vue" href="/docs/guides/vue/vite" />
+  <Card title="Nuxt" href="/docs/guides/vue/nuxt" />
+  <Card title="Webflow" href="/docs/guides/webflow" />
+  <Card title="Wix" href="/docs/guides/wix" />
+  <Card title="WooCommerce" href="/docs/guides/woocommerce" />
+  <Card title="WordPress" href="/docs/guides/wordpress" />
 </Cards>
+
+## Code
+
+Rybbit is open sourced under the AGPL-3.0 license. You can find the code on [GitHub](https://github.com/rybbit-io/rybbit).
+
+## Community
+
+Join our [Discord server](https://discord.gg/DEhGb4hYBj) or talk to me on [X](https://x.com/yang_frog). I'm always looking for feedback and ideas to make Rybbit as awesome as possible.
