+++ docs-v2/content/docs/inviting-users.mdx
@@ -1,0 +1,39 @@
+---
+title: Inviting Users
+description: You can invite your team members to your Rybbit organization as either an admin or a member
+---
+
+import { Steps, Step } from 'fumadocs-ui/components/steps';
+import { Callout } from 'fumadocs-ui/components/callout';
+
+<Callout type="info">
+  On self-hosted instances, this workflow is slightly different. Instead of inviting users, you can create a new user and set their password directly since self-hosted instances don't send emails. 
+</Callout>
+
+<Steps>
+<Step>
+### Add the user to your organization
+
+Go to https://app.rybbit.io/organization "Add Member" button in the top right. Enter the user's email address and select the user's role.
+</Step>
+
+<Step>
+### Select the user's role
+ 
+You can either make the user an admin or a member. 
+
+**Admins** can invite other users, add websites, edit websites, delete websites, and see all website data.
+
+**Members** can see all website data, but they cannot edit anything.
+</Step>
+
+<Step>
+### Send the invite
+
+Click the "Invite" button. The user will receive an email with a link to sign up for Rybbit. You can manage all invites in the organizations tab.
+</Step>
+</Steps>
+
+<Callout type="info">
+  Currently, each user can only be in one organization.
+</Callout>
