+++ docs-v2/content/docs/meta.json
@@ -1,0 +1,32 @@
+{
+  "title": "Documentation",
+  "pages": [
+    "index",
+    "script",
+    "track-events",
+    "api",
+    "definitions",
+    {
+      "title": "Guides",
+      "pages": ["guides"]
+    },
+    {
+      "title": "SDKs",
+      "pages": ["sdks"]
+    },
+    {
+      "title": "Self-Hosting",
+      "pages": ["self-hosting", "self-hosting-manual", "self-hosting-advanced", "self-hosting-guides"]
+    },
+    "hiding-own-traffic",
+    "inviting-users",
+    "localhost-tracking",
+    "public-site",
+    "changing-domains",
+    "deleting-sites",
+    "enhanced-privacy",
+    "troubleshooting",
+    "v1-migration",
+    "roadmap"
+  ]
+}
