+++ docs-v2/content/docs/public-site.mdx
@@ -1,0 +1,16 @@
+---
+title: Make your analytics public
+description: Learn how to make your analytics dashboard public and share it with others
+---
+
+import Image from "next/image";
+
+# Make your analytics public
+
+You can make your analytics public by going to the settings page and enabling the "Public Analytics" option. You can then share the page link with anyone you want. 
+
+- External viewers will not be able to edit any of your settings or add/remove/edit any existing reports, funnels, or goals. 
+- Your other websites will not be affected.
+
+<br/>
+<Image src="/settings.jpg" alt="Public Site" width={500} height={500} />
