+++ docs-v2/content/docs/roadmap.mdx
@@ -1,0 +1,24 @@
+---
+title: Roadmap
+description: Planned features and upcoming development for Rybbit analytics platform
+---
+
+# Roadmap
+
+Rybbit is a quickly growing project. We already have quite a few features, but there is much more to come.
+
+## Planned features
+
+- Data export
+- Data import from Google Analytics, Plausible, Umami, PostHog, and more
+- Custom reports/dashboards
+- Stats API
+- More tracking script/SDK integration guides
+- Mobile app support
+- Custom themes
+- Revenue tracking
+- Email reports
+- Traffic filters/bot blocking
+- Traffic alerts
+
+Once again, join our [Discord server](https://discord.gg/DEhGb4hYBj) or follow me on [X](https://x.com/yang_frog) to share your feedback and ideas. We would love to hear from you!
