+++ docs-v2/content/docs/sdks/meta.json
@@ -1,0 +1,7 @@
+{
+  "title": "SDKs",
+  "pages": [
+    "web",
+    "node"
+  ]
+}
