+++ docs-v2/content/docs/self-hosting-guides/meta.json
@@ -1,0 +1,7 @@
+{
+  "title": "Self-Hosting Guides",
+  "pages": [
+    "custom-nginx",
+    "nginx-proxy-manager"
+  ]
+}
