+++ docs-v2/content/docs/self-hosting-manual.mdx
@@ -1,0 +1,296 @@
+---
+title: Manual Docker Compose Setup
+description: Manual setup guide for Rybbit using Docker Compose without the setup script
+---
+
+import { Callout } from 'fumadocs-ui/components/callout';
+import { Steps, Step } from 'fumadocs-ui/components/steps';
+import { Tabs, Tab } from 'fumadocs-ui/components/tabs';
+
+# Manual Docker Compose Setup
+
+This guide is for users who want to manually set up Rybbit using Docker Compose without the setup script. This is useful if you already have a reverse proxy setup (like Nginx, Traefik, or Coolify), or if you want more control over the configuration.
+
+## Prerequisites
+
+- Docker and Docker Compose installed
+- A domain name pointed to your server
+- Basic knowledge of Docker and environment variables
+
+## Setup Steps
+
+<Steps>
+<Step>
+### Clone the Repository
+
+```bash
+git clone https://github.com/rybbit-io/rybbit.git
+cd rybbit
+```
+</Step>
+
+<Step>
+### Create Environment File
+
+Create a `.env` file in the root directory with the following content:
+
+```bash
+# Required: Your domain and base URL
+DOMAIN_NAME=your.domain.com
+BASE_URL=https://your.domain.com
+
+# Required: Authentication secret (generate a random 32+ character string)
+BETTER_AUTH_SECRET=your-very-long-random-secret-string-here
+
+# Optional: Disable new user signups after creating admin account
+DISABLE_SIGNUP=false
+
+# Optional: Custom ports (only needed if you want different ports)
+# HOST_BACKEND_PORT="3001:3001"
+# HOST_CLIENT_PORT="3002:3002"
+
+# Optional: Database credentials (defaults work fine)
+# CLICKHOUSE_PASSWORD=frog
+# POSTGRES_USER=frog
+# POSTGRES_PASSWORD=frog
+# POSTGRES_DB=analytics
+# CLICKHOUSE_DB=analytics
+```
+
+<Callout type="info">
+To generate a secure `BETTER_AUTH_SECRET`, you can use:
+```bash
+openssl rand -hex 32
+```
+</Callout>
+</Step>
+
+<Step>
+### Choose Your Setup Method
+
+<Tabs items={['Built-in Caddy (Automatic SSL)', 'Without Caddy']}>
+<Tab value="Built-in Caddy (Automatic SSL)">
+If you don't have a reverse proxy and want automatic SSL certificates, use the default docker-compose.yml:
+
+```bash
+docker compose up -d
+```
+
+This will:
+- Start all services including Caddy
+- Automatically obtain SSL certificates
+- Make your app available at `https://your.domain.com`
+</Tab>
+<Tab value="Without Caddy">
+If you have your own reverse proxy (Nginx, Traefik, Coolify, etc.), you need to exclude Caddy and expose the ports:
+
+1. **Option B1: Modify the .env file**
+   ```env
+   # Add these lines to expose ports
+   HOST_BACKEND_PORT="3001:3001"
+   HOST_CLIENT_PORT="3002:3002"
+   ```
+
+   Then start without Caddy:
+   ```bash
+   docker compose up -d backend client clickhouse postgres
+   ```
+
+2. **Option B2: Use docker-compose.override.yml**
+   Create a `docker-compose.override.yml` file:
+   ```yaml
+   services:
+     backend:
+       ports:
+         - "3001:3001"
+     client:
+       ports:
+         - "3002:3002"
+   ```
+
+   Then start without Caddy:
+   ```bash
+   docker compose up -d backend client clickhouse postgres
+   ```
+</Tab>
+</Tabs>
+</Step>
+
+<Step>
+### Configure Your Reverse Proxy
+
+If you're using your own reverse proxy, configure it to:
+- Proxy requests to `/api/*` to `http://localhost:3001`
+- Proxy all other requests to `http://localhost:3002`
+
+Example configurations:
+
+<Tabs items={['Nginx', 'Traefik']}>
+<Tab value="Nginx">
+```nginx
+server {
+    listen 80;
+    server_name your.domain.com;
+    return 301 https://$host$request_uri;
+}
+
+server {
+    listen 443 ssl;
+    server_name your.domain.com;
+
+    # Your SSL configuration here
+
+    location /api/ {
+        proxy_pass http://localhost:3001;
+        proxy_set_header Host $host;
+        proxy_set_header X-Real-IP $remote_addr;
+        proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+        proxy_set_header X-Forwarded-Proto $scheme;
+    }
+
+    location / {
+        proxy_pass http://localhost:3002;
+        proxy_set_header Host $host;
+        proxy_set_header X-Real-IP $remote_addr;
+        proxy_set_header X-Forwarded-For $proxy_add_x_forwarded_for;
+        proxy_set_header X-Forwarded-Proto $scheme;
+    }
+}
+```
+</Tab>
+<Tab value="Traefik">
+```yml
+services:
+  backend:
+    labels:
+      - "traefik.enable=true"
+      - "traefik.http.routers.rybbit-api.rule=Host(`your.domain.com`) && PathPrefix(`/api`)"
+      - "traefik.http.routers.rybbit-api.tls.certresolver=letsencrypt"
+      - "traefik.http.services.rybbit-api.loadbalancer.server.port=3001"
+  
+  client:
+    labels:
+      - "traefik.enable=true"
+      - "traefik.http.routers.rybbit-client.rule=Host(`your.domain.com`)"
+      - "traefik.http.routers.rybbit-client.tls.certresolver=letsencrypt"
+      - "traefik.http.services.rybbit-client.loadbalancer.server.port=3002"
+```
+</Tab>
+</Tabs>
+</Step>
+
+<Step>
+### Start the Services
+
+```bash
+# Start all services
+docker compose up -d
+
+# Or start specific services (without Caddy)
+docker compose up -d backend client clickhouse postgres
+```
+</Step>
+
+<Step>
+### Verify Setup
+
+Check that all services are running:
+```bash
+docker compose ps
+```
+
+Monitor logs:
+```bash
+docker compose logs -f
+```
+</Step>
+
+<Step>
+### Create Admin Account
+
+Navigate to `https://your.domain.com/signup` and create your admin account.
+</Step>
+</Steps>
+
+## Service Architecture
+
+Rybbit consists of these services:
+
+- **client**: Next.js frontend (port 3002)
+- **backend**: Node.js API server (port 3001)
+- **postgres**: User data and configuration
+- **clickhouse**: Analytics data storage
+- **caddy**: Reverse proxy with automatic SSL (optional)
+
+## Common Configurations
+
+### Using Custom Ports
+
+If you need different host ports (e.g., if 3001/3002 are already in use):
+
+```env
+HOST_BACKEND_PORT="8080:3001"
+HOST_CLIENT_PORT="8081:3002"
+```
+
+### Using with Coolify
+
+In Coolify, you can:
+1. Create a new service with the docker-compose.yml
+2. Set the environment variables in Coolify's interface
+3. Use Coolify's built-in proxy by exposing ports 3001 and 3002
+
+### Using with Nginx Proxy Manager
+
+1. Set up port exposure in your .env:
+   ```env
+   HOST_BACKEND_PORT="3001:3001"
+   HOST_CLIENT_PORT="3002:3002"
+   ```
+
+2. Create a proxy host pointing to your server IP:3002
+3. Add a custom location `/api/*` pointing to your server IP:3001
+
+### Database Persistence
+
+Data is automatically persisted in Docker volumes:
+- `rybbit_clickhouse-data`: Analytics data
+- `rybbit_postgres-data`: User accounts and site configurations
+
+## Advanced Configurations
+
+For specific reverse proxy setups, see our detailed guides:
+
+- **[Custom Nginx Setup](/docs/self-hosting-guides/custom-nginx)** - Complete Nginx configuration with SSL
+- **[Nginx Proxy Manager](/docs/self-hosting-guides/nginx-proxy-manager)** - NPM setup without port exposure
+
+These guides provide step-by-step instructions for integrating Rybbit with your existing infrastructure.
+
+## Troubleshooting
+
+### Services won't start
+```bash
+# Check logs
+docker compose logs
+
+# Check service status
+docker compose ps
+```
+
+### Port conflicts
+If you get port binding errors, either:
+1. Change the host ports in your .env file
+2. Stop the conflicting service
+3. Use docker-compose.override.yml to customize ports
+
+### Database connection issues
+Ensure the database services are healthy:
+```bash
+docker compose ps
+```
+
+Both postgres and clickhouse should show "healthy" status.
+
+<Callout type="info">
+For more advanced configurations and setup script options, see our [Advanced Self-Hosting Guide](/docs/self-hosting-advanced).
+</Callout>
